requirements.txt
python-docx
pywin32 ; platform_system == "Windows"